requirements.txt
asgiref==3.2.10
boto3==1.17.36
botocore==1.20.36
certifi==2020.12.5
chardet==4.0.0
Django==3.1
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-session-timeout==0.1.0
django-storages==1.11.1
dj-database-url==0.5.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Pillow==7.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.1
python-decouple==3.4
pytz==2020.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.3
reportlab==4.4.0